feat: work on global function and classes organization
