+++ src/new-account-list.js
@@ -1,0 +1,40 @@
+class AccountList {
+  constructor() {
+  }
+}
+
+class DeleteAccountList extends AccountList {
+  constructor() {
+    super();
+  }
+
+  createListItem() {
+
+  }
+
+  createList() {
+
+  }
+
+  resetList() {
+
+  }
+}
+
+class PasteAccountList extends AccountList {
+  constructor() {
+    super();
+  }
+
+  createListItem() {
+
+  }
+
+  createList() {
+
+  }
+
+  resetList() {
+
+  }
+}
